Finish the performance optimization
I wrapped up the optimization I was working on. I got some pretty decent performance gains out of one particular function that was taking most of the CPU time (The stat generation) for the test I was running.

The optimization involves precalculating the combined prefix/suffix weights so I don't need to repeatedly sum them when they are just going to always be a static list.