Crafting Income
The results page now displays the total income (or loss) from the crafting process.

The user can also now enter the cost of the item base. This cost is factored into the final income. Scouring orbs will automatically be used on failed items if the cost of the item base is greater than that of a scouring orb and the failed item is not corrupted.

A more detailed breakdown of what currency is used and when it's used will be included in a future feature.

Several bugs have also been fixed.